refactor(dsl): split GraphBuilder/factories/executor into single-responsibility modules
Pure structural refactor with no behavioural changes. Breaks the three
large DSL files in `src/main/dsl/` into per-class / per-concern files
under new subdirectories. Each file is <250 LOC and has a single clear
responsibility. The original three files become barrel re-exports so
callers (main.ts, span analyzers, tests) keep working unchanged.

Splits:
- `GraphBuilder.ts` (1527 LOC) -> `graph/` (15 files: types, cartesian,
  signalResolution, moduleOutput, moduleOutputWithRange, deferredOutput,
  collection, collectionWithRange, deferredCollection, bus, moduleNode,
  builder, buildOutputs, serialize, index)
- `factories.ts` (424 LOC) -> `factory/` (7 files: spanRegistry,
  identifiers, namespaceTree, createFactory, dslContext, units, index)
- `executor.ts` (635 LOC) -> `executor/` (7 files: types, arrayPipe,
  buildContext, runFunction, executePatchScript, validateDSLSyntax,
  index) plus `composites/` (6 files for $deferred, $bus, $buffer,
  $delay, $slider, settings), `wavs/proxy.ts`, and `table/$table.ts`.

Verification: 170/170 unit tests pass. No imports outside `src/main/dsl/`
change. Sets up the relocate to `@modular/dsl` in PR 3.
