+++ src/main/dsl/executor/index.ts
@@ -1,0 +1,7 @@
+export { executePatchScript } from './executePatchScript';
+export { validateDSLSyntax } from './validateDSLSyntax';
+export type {
+    DSLExecutionResult,
+    DSLExecutionOptions,
+    WavsFolderNode,
+} from './types';
